Wiki/Rockets: load rockets

+++ app/src/main/java/com/github/emilg1101/spacex/data/spacexapi/pojo/response/Mass.kt
@@ -3,6 +3,6 @@ package com.github.emilg1101.spacex.data.spacexapi.pojo.response
 import com.google.gson.annotations.SerializedName
 
 data class Mass(
-    @SerializedName("kg") val kg: Int,
-    @SerializedName("lb") val lb: Int
+    @SerializedName("kg") val kg: Double,
+    @SerializedName("lb") val lb: Double
 )
